adov/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
firebase-admin>=6.6.0
anthropic>=0.40.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
sse-starlette>=2.1.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
requests>=2.31.0
apify-client>=1.6.0
Pillow>=10.0.0
rapidocr-onnxruntime>=1.3.0
openai>=1.0.0
PyYAML>=6.0
